+++ NEWS.md
@@ -117,7 +117,13 @@
  * Fix bug with `sampleInstances=FALSE` that could re-evaluate the same
    (instance, seed) pair more than once.
 
-
+ * Fix bug in `ablation_cmdline()` about missing `scenario` object.
+ 
+ * `ablation()` will now save and restore the previous random seed.
+ 
+ * `ablation()` will detect if the logfile (e.g., `irace.Data`) is incomplete.
+ 
+ 
 # irace 3.5
 
 ## New features and improvements